Added file of dbs decomposition
This is to make the creation of tables easier, forgot to finish the MER, but in essence the idea is that the table laptop and prices are connected by the primary key of laptops which is lid. (This information is possible as Conrado is taking CS334). Tried doing a 3NF decomposition, came to the conclusion that with our dataset that is as close as we can get to it before we make a lossy join.