Added possible Decomposition file
The idea behind this file was that Conrado tried to do a BCNF decomposition of the dataset and decided that with the way that the dataset is set up it makes no sense to do so as otherwise each query, or insertion, or deletion, etc. Would take an enormous amount of time and would add a lot of complexity to it, which is not necessary or desired.